api/requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.3
psycopg2
flask-json-schema
cassandra-driver
marshmallow